Add files via upload
Campaign: Enter the name of the campaign and it will show you the results for all the creatives running within that campaign. This is helpful in understanding which creative(s) is/are performing well versus the rest
By Ad and Adset: Most of the campaigns have creatives info passed to us under af_ad while there are some that have it coming under af_adset. You can look for a given creatives performance by typing out the name or a keyword of a given creative and the program will return the results for that creative across all the campaigns where it's active. This option is helpful if there's a creative which is running across different campaigns/networks/platforms and you want to learn how it's been performing across the board